+++ data/src/test/java/de/ph1b/audiobook/data/ChapterFactory.kt
@@ -10,7 +10,7 @@ internal object ChapterFactory {
   fun create(
     file: String = "First.mp3",
     parent: String = "/root/",
-    duration: Int = 100,
+    duration: Long = 100,
     lastModified: Long = 12345L,
     bookId: UUID,
     marks: SparseArrayCompat<String> = emptySparseArray()
